requirements.txt
python-dotenv
python-telegram-bot
openai
python-telegram-bot[job-queue]